AUT-4248: Set activeMfaMethodId on session initialisation
When a user visits the /authorize path they may already have a session with some user session data.

We previously carried over their email and phoneNumber. For method management work we moved phoneNumber into mfaMethods.

In recent commits we set an activeMfaMethodId to indicate which mfaMethods item is in use for the journey (a user can now change those).

However, we need to also set it here when we uplift a user, so when they go straight to the /enter-code screen we can render the page correctly (by knowing which MFA to use)