AUT-4272: skip csrf protection on all routes
Note that this will be re-enabled as part of the third PR under this ticket. We want to ensure the csrf token is being set on the session and injected into the form hidden input before we start enforcing protection. This is with the aim of ensuring partially completed user journeys (with an active session) are not affected during the deployment (which would otherwise fail due to a "csrf error", e.g. if no token was in the hidden form input at the time of submit).