BAU: Simplify dev orchstub proxy
Previously our `dev-app-orchstub.js` would make a request to the orchstub and forward on the returned payload and cookies. When using this against a hosted orchstub and localhost frontend there would be a mismatch between the `gs` cookies set on the domains.

Here we utilise a new orchstub feature to allow the journey to start on a GET request, allowing successful setup of the same session across hostnames.