AUT-4347: add tests for expected string given subtheme
This file doesn't have a lot of coverage at the moment. This provides some level for the changes I've made in this commit, as coverage was <80% on this PR. The test isn't very complex, but I think it provides a means to guarantee that the expected entry from translations is being used, so I think it has value, even if it's simple. If we like it, we should expand to fully test this validation file.