AUT-4280: Cookie banner should not be displayed for the GOV.UK App
Cookies should be rejected by default like for the V2 App.

Includes changes to tests which check the cookie consent behaviour is correct for the channel.

Previously the test succeeded regardless of whether the channel should present consent or not, because the 'createMockCookieConsentService' created a mock which always returned a response with the consent value passed in when creating the mock, rather than the value passed in when using the mock.  A new mock has been created which responds to the values of paramaters passed in, see 'createCookieConsentMock'.