AUT-4452: Fix missing journey for SMS OTP requests
We have a bug on password reset journeys where when a user triggers resending an SMS OTP it was being misattributed to the SIGN_IN journey, not PASSWORD_RESET_MFA. This meant that the count was also assigned to SIGN_IN incorrectly.

Here this is corrected by ensuring when requesting an SMS OTP via the /mfa API we send the correct journey.

This bug was also apparent when an SMS was sent after a user chooses their backup SMS MFA method.