AUT-4452: Fix SMS OTP send trigger on password reset
We have a bug where a user will trigger two SMS OTP sends when attempting to try a resend on the password reset journey.

This happened as we were triggering an SMS OTP send on both the loading GET /reset-password-2fa-sms and POST /resend-code, and the "POST" was followed by a redirect to the "GET", thus triggering a second SMS send.

Now instead of sending an SMS on the GET /reset-password-2fa-sms, this is moved to the POST /reset-password-check-email.

i.e. user submits their email OTP on a password reset journey. If correct an SMS OTP is sent (if their DEFAULT MFA is SMS). If a user tries to resend, they follow links and trigger POST /resend-code, which redirects back to GET /reset-password-2fa-sms but crucially this redirect does not trigger another SMS.