AUT-4452: Check MFA lockout just after password entry
Now that we have merged counts so they are MFAMethodType agnostic, we need to ensure that when a BACKUP SMS MFA usage triggers a lockout, the user is not permitted to restart their journey using their DEFAULT AUTH_APP MFA.

They can do so currently. This is because we do not check after the user has entered their password for any MFA lockouts before redirecting to the AUTH_APP entry page. If their DEFAULT was an SMS MFA, then the API call to send the SMS OTP would return the error instead.

Here we add the MFA check to the POST /enter-password route. If the /login API returns an MFA lockout error code we instead render the error message straight away, rather than allowing them to proceed to the MFA OTP entry page.