BAU: Fix multiple nodes being highlighted when clicking on a state
We nodes have a trailing "-[NUMBER]" so we just did a startsWith to find nodes to highlight. However, this meant that nodes that started with the same string were both highlighted.

Now, the trailing number is removed and the strings are strictly compared, so we only highlight the node that was actually clicked.