AUT-4563: 400 when provided session ID invalid
We get the sessionId from the `gs` cookie at the moment. We break the value up in the session-middleware.ts to get the sessionID and clientSessionId.

When the cookie is not provided, those values are not set in req.locals.

We then attempt to send those values to the backend, but if they are not provided then the backend returns error 1000, `SESSION_ID_MISSING_OR_INVALID`.

This is a client side error as far as authentication is concerned. The client should provide the cookie, and it did not.

So instead of returning an internal server error 500 status code, return a 400 which is more appropriate.

We are still using the exsiting 500 error template to render to the user while we await new content to show.