AUT-4644: Host apple-app-site-association file in non-prod
The mobile team would like to experiment with a link between the site and app with the help of the /.well-known/apple-app-site-association file.

This initial implementation hosts the provided content in all envs apart from production and integration.

Depending on how the experiment goes, and how future file content changes, we can change the implementation to suite.

For now, this is a minimal change to get the file working initially.