BAU: Remove old local  RP stubs
The old local RP stubs relied on calling the orch authorize endpoint in order to create a backend session for the local frontend.

Now that orch is no longer deployed to the auth dev environment, and it's no longer possible to connect to a remote api in build and staging as the apis are now private, these stubs no longer work so must be removed in favour of the local orch stub.