AUT-3933: Add progress button template and translations
The way the `@govuk-one-login/frontend-ui` package implements translations is different to our approach. We have opted to follow our existing approach for sourcing translations.

Note that a template has been used over a macro as the context is not available within the macro for the translations to be performed correctly. This is the same strategy used by the Google Analytics (ga4) component.

Usage of the new progress-button wrapper template should be roughly as follows (assuming `scriptNonce` is an available parameter):
```
    {% set progressButtonParams = { nonce: scriptNonce } %}
    {%- include "common/progress-button/template.njk" -%}
```